[js_fuzzer] Add wasm-only flavor
This adds a new runner flavor that only loads test cases from the
V8 and Fuzzilli corpora that depend on the wasm module builder.

Mutations in this runner are slightly tweaked to be more
"wasm friendly".

Existing fuzzing behavior is not affected as the implementation is
contained to new corpus and runner subclasses.

We hope that running in this mode might produce more interesting
wasm test cases.

Bug: 400854636
Change-Id: I96b6155b0cf5c91eca637c42ccd7a58b1dafd31a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6328639
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Carl Smith <cffsmith@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99089}