[wasm] Use cppgc for ParamBuffer of the generic js-to-wasm wrapper
The generic js-to-wasm wrapper needs a byte array called ParamBuffer to
store parameters that have already been converted from the JS domain to
the WebAssembly domain. So far this buffer was malloc'ed in C++, which
requires manual deallocation. With this CL the buffer gets allocated on
the cppgc heap so that it gets deallocated by the GC. This should fix a
memory leak that was found by clusterfuzz.

Since with this CL, cppgc references can be on the JS heap, the JS heap
has to be scanned by conservative stack scanning. This was not done so
far in simulator builds, and therefore this CL also adds support for
scanning the simulator stack for references.

Bug: 391659743
Change-Id: I66b1d1de65238ef0e2d472ebda95e9a8f1da9257
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6281089
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Omer Katz <omerkatz@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99200}