[turboshaft][arm64] Fix PGO instrumentation compilation
In general, we normalize loads and stores in Turboshaft before they
reach the instruction selector (this is done by
LoadStoreSimplificationReducer). However, this runs before
BlockInstrumentationReducer (which introduces the instrumentation
required to collect PGO data), and this reducer introduces loads &
stores that are not normalized, leading to CHECK failures during
instruction selection.

So, this CL adds LoadStoreSimplificationReducer to the
BlockInstrumentationReducer, so that all stores/loads are correctly
normalized.

Fixed: 391722207
Change-Id: I22592b9c5e49666de6578aa131f121bd7fae68fe
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6330139
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Auto-Submit: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Nico Hartmann <nicohartmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99097}