[base] Drop the lock during decommit in BoundedPageAllocator
From profiles, we see that contention on this lock takes a non-trivial
amount of time in speedometer. Make sure we don't hold it for too
long. This may not improve toplevel metrics though, since there is
locking at the lower levels (in the page allocator, or the kernel).

Change-Id: I6bd06a500b51487be1af83c2e9811be25faf82e2
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6332712
Reviewed-by: Anton Bikineev <bikineev@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Benoit Lize <lizeb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99115}