[js_fuzzer] Fix loading dependencies from the Fuzzilli corpus
Loading V8 dependencies from Fuzzilli, e.g. the wasm module builder,
was not yet fixed correctly. Tests unfortunately didn't catch this.

Concretely all dependencies are loaded in the context of a
particular corpus. This triggers a corpus-specific way to load flags.
When trying to load the wasm module builder (from V8) in the
context of the Fuzzilli corpus, the Fuzzilli flag-loading logic
was triggered and ran into an assert about the missing settings.json.

This fix adds corpus-specific dependency-loading logic and requires
the Fuzzilli corpus to point to a side-by-side V8 corpus to resolve
V8 dependencies. The places where we initialize a Fuzzilli corpus
are updated, so are the tests.

Patch 1 improves an existing test, so that it reproduces the problem:
https://crrev.com/c/6333173/1

Patch 5 is the fix:
https://crrev.com/c/6333173/2..5

Bug: 400854636
Change-Id: I80aa0d45afb647d4e0f7c572804233b9113797b9
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6333173
Reviewed-by: Carl Smith <cffsmith@chromium.org>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99108}