Fix test262 case for arg empty case to follow sec-regexp.p-@@replace
According to the RegExp.prototype.@@replace spec [1], the 'captures' should be constructed from the start index 1:
    h. Let n be 1.
    i. Repeat, while n ≤ nCaptures,
       i. Let capN be ? Get(result, ! ToString(𝔽(n))).
       ii. If capN is not undefined, then
           1. Set capN to ? ToString(capN).
       iii. Append capN to captures.

And the arguments should be passed with concated with 'matched':

    k. If functionalReplace is true, then
       i. Let replacerArgs be the list-concatenation of « matched », captures, and « 𝔽(position), S ».

However, the current implementation iterates from 0 and does not concatenate with first matched. As a result, the test case 'fn-invoke-args-empty-result' has been failed.

Hence, fix the behavior to intent with spec for replace by placing the first match element in call_args.

[1] https://tc39.es/ecma262/#sec-regexp.prototype-%symbol.replace%



Bug: 42200389
Change-Id: Ie8bcac7a8925baa389ba74ef7aaf3361e2a6556d
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6317726
Commit-Queue: Seokho Song <seokho@chromium.org>
Reviewed-by: Rezvan Mahdavi Hezaveh <rezvan@chromium.org>
Reviewed-by: Shu-yu Guo <syg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99166}