[wasm][jspi] Fix simulator stack underflow
To perform runtime calls with different signatures, some simulators
prepare a fixed number of arguments which is an upper bound of the
actual parameter count. The extra stack slots contain arbitrary data and
are never used, but with stack-switching this can happen close to the
stack start, so we need to reserve a safety gap to ensure that the
addresses are at least mapped to prevent a crash.

R=jkummerow@chromium.org

Fixed: 400996817
Change-Id: I8222ad09da70ae2e11cef709d6eac651c7462982
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6330308
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99128}