[wasm][jspi] Remove jump buffer external pointer
Remove the jump buffer external pointer and load it from the stack
memory instead. This was initially introduced as a small optimization
to avoid one indirection, but the benefit is not clear now that it has
to go through the EPT anyway. In addition, it can be corrupted and
increases the sandbox attack surface.

R=jkummerow@chromium.org

Bug: 384549659
Change-Id: I847a711f5cd2fd30a63cac7c4ed78c16aa340bf3
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6276515
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99155}