[regalloc] Cleanup a DCHECK
Follow-up to:
https://chromium-review.googlesource.com/c/v8/v8/+/6320155

A better fix to the original issue is to only apply the DCHECK for fixed
inputs and fixed outputs, but not fixed temporaries. Non-allocatable
fixed temporaries can be accepted since they don't interfere with other
register allocation decisions.
This avoids the RISCV-specific code, and strengthens the check on that
platform.

R=jkummerow@chromium.org

Change-Id: I84c903080358147bb790adbdb486537e7f31c53e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6336030
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99161}