Fix data race in backing store flags
The flags are stored in a C++-defined bitfield consisting of 9 bits
(i.e. stored in two bytes). During serialization we write the
`globally_registered_` bit under a mutex, but we also read this bit and
others without holding the mutex.

While this is fine, it triggers TSan reports. So this CL refactors this
bitfield into a `base::EnumSet`, which is wrapped in a `std::atomic`
which allows to read it concurrently and update it via CMPXCHG.

R=ishell@chromium.org

Fixed: 401633689
Change-Id: If1f2d1a7fd703fbb0b88dfc2a70cbf148108ce8e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6336031
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99171}