Make isolate id accessible v8 internally.
We have isolate id stored in the inspector, but we'd like to make
isolate id accessible outside of inspector so instead we're storing the
isolate id in the debug instance which has a 1-1 relationship with
isolate. The isolate id can help enhanced traces uniquely identify a
script as script can have the same id across different isolates.

Bug: 337909145
Change-Id: Iced8e89cc4dd7a8a23383960502735dd7c75172f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6324757
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Jayson Chen <jaysonchen@microsoft.com>
Reviewed-by: Simon Zünd <szuend@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99170}