[wasm] Implement memory control's unmap
This CL adds an unmap method as a method that makes sure that all memory
pages that got mapped in originally by a map() call are not mapped
anymore. `unmap()` does that by mapping new, zeroed pages over the pages
of the original map() call.

Since `unmap()` is a method on the WasmMemoryMapDescriptor, this CL also
makes `map()` a method the the WasmMemoryMapDescriptor, instead of a
method on the WasmMemoryObject.

R=dlehmann@chromium.org

Bug: 339678654
Change-Id: Iadf5e6ac27fd49f8f496ced0349fac08cca56a9f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6309162
Reviewed-by: Daniel Lehmann <dlehmann@chromium.org>
Reviewed-by: Clemens Backes <clemensb@chromium.org>
Commit-Queue: Andreas Haas <ahaas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99199}