[heap] Do not reset allocation limit with initial old gen size
This CL is a prerequisite for finching an initial old gen size with
V8. It stops resetting the limit to the the initial old gen size
if one was explicitly passed as argument on page navigations. This
is an alternative mode to the "unconfigured mode" before the first
full GC, where we start with an allocation limit of 512M and later
shrink on scavenges.

When using 32M as initial old gen size, this change (when enabled)
is performance-neutral on Speedometer and Jetstream. However, it
removes a few full GCs with very small heaps in the beginning of
the benchmarks.

Bug: 391129951
Change-Id: If3a8fd411c8c7e1405f62dec0fde599bb3c4bb6c
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6333085
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99213}