[maglev-inlining] Refactor MaglevCallerDetails
This tries to re-use information already in the call node.

It delays the creation of the inner MaglevCompilationUnit
and avoids re-creating the lazy deopt frame in case of
non-eager calls.

It also renames VirtualObject::List to VirtualObjectList, so
that we can forward declare the class.

Bug: 393989628
Change-Id: I8ee5ac42820f741de632b02c3a977b1f753fb815
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6348596
Reviewed-by: Olivier Flückiger <olivf@chromium.org>
Auto-Submit: Victor Gomes <victorgomes@chromium.org>
Reviewed-by: Nico Hartmann <nicohartmann@chromium.org>
Commit-Queue: Victor Gomes <victorgomes@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99229}