[wasm] Fix streaming instantiation with manipulated Promise
The error is very similar to https://crbug.com/374820218, but the fix
that was applied for streaming compilation was missing from the
streaming instantiation path.
Since both share a lot of code, this CL introduces a common helper which
has the check for failed execution of the `Promise.then` step.

R=jkummerow@chromium.org

Fixed: 396461004
Change-Id: Idf5d352041d0792f7b0966607fd1ff5b0782ffe4
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6351696
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99239}