[compiler][arm64] Let jump-threading elide bti instructions
Table switches use indirect branches and so each case needs a `bti`
instruction in order to conform with CFI. The instruction is encoded as
a property of the block, as a switch target, and gets propagated by jump
threading. However, if jump threading omits a block, the property needs
to be dropped otherwise an extra `bti` instruction will be generated and
the block will not be fully removed.

Change-Id: Ica1a92ef17e37c62889acb1d0a9b9df88e351f34
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6352143
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Pierre Langlois <pierre.langlois@arm.com>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99304}