[heap] Don't abort compaction on the same page twice
During evacuation we check that the number of pages we aborted
compaction for matches the number of evacuation candidates found that were aborted.
If a page is pinned in the full GC, we will abort compaction for it.
However we can try to abort compaction for it again later (e.g. in a
GC with stack if the GC happens during a fast C call or if some flags
are set). In that case, we will abort compaction again, count the page
twice, and the above mentioned evacuation will fail.

Bug: 403340573
Change-Id: I8fa1f1eac543d915332bef0eef82d336d82c4756
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6356828
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Omer Katz <omerkatz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99258}