[json] Stringify: Use union in ContinuationRecord
[1] added a couple of fields to the ContinuationRecord that are only
relevant for JSObjects, but in turn JSObjects index/length can be stored
in a uint16_t (whereas we need uint32_t for JSArrays).

So with this CL we store fields per ContinuationType in a union. We
also turn ContinuationRecord into a proper class with accessors and
static factory methods to DCHECK that all stores/load are appropriate.

[1] https://crrev.com/c/6343190

Bug: 380044242
Change-Id: I0d4e8991561896d7f34b3b43b85a0196ae29f02c
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6351924
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Patrick Thier <pthier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99266}