[turboshaft] Remove LoopPeelingPhase
Loop peeling was never meant to run in the backend, but was added to
the backend (disabled though), in order to get some fuzzer coverage
until we'd actually need it.

Except that by now, it's clear that we'll never need it, since loop
peeling should happen in the frontend, and Turboshaft will not be used
for the frontend: we'll use Maglev instead.

Fixed: 403211343
Change-Id: I6e2a3ad56895cb9e73e0ca7fe42adb621a1b8450
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6360437
Auto-Submit: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Matthias Liedtke <mliedtke@chromium.org>
Reviewed-by: Matthias Liedtke <mliedtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99282}