[maglev] Refactor try-catch mechanism
This is based on the realization that necessary
MergeThrows happen in the same graph builder. We do not
need to MergeThrow inside an inlined function, therefore
we do not need to keep track of the graph builder that was
used to create the merge.

This refactor will allow us to support try-catch for
non-eager inlinin, which is currently blocked by
https://crbug.com/403345124

Bug: 393989628
Change-Id: Iac882b4bf40f55c36c02b4d61448f747cd5bf75c
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6355017
Commit-Queue: Victor Gomes <victorgomes@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99285}