[js_fuzzer] Validate cross-over insertions
This adds a fuzzing mode that deterministically enumerates all
expressions in the snippet DB and inserts each into a predefined
template. The resulting test cases can be used to test the
quality of the resulting code (to be done in a follow up).

Patch 1 adds a small refactoring that doesn't change behavior.
One method in the cross-over mutator is extracted to be reused.
https://crrev.com/c/6357589/1

Patch 5 is the implementation:
https://crrev.com/c/6357589/1..5

Bug: 389069288
Change-Id: If3d7720a9902cd5fa1af1ef762e385a5da4b9f24
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6357589
Reviewed-by: Carl Smith <cffsmith@chromium.org>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99295}