[test] Add a test variant that stresses maglev tests with turbofan
This adds a flag that makes %OptimizeMaglevOnNextCall optimize to
Turbofan, i.e. the reverse of the existing flag:
--optimize-on-next-call-optimizes-to-maglev

This would have detected https://crbug.com/403345130 earlier,
which had a regression test for Maglev that also exposed a
Turbofan bug.

We run this new variant only with Mjsunit on the most important
bots.

Bug: 403345130
Change-Id: I2cc3ba975c8247767c98ed0d26af80a235dace3a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6360775
Reviewed-by: Marja Hölttä <marja@chromium.org>
Reviewed-by: Darius Mercadier <dmercadier@chromium.org>
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99309}