[test] Add DB e2e test
This adds a more extensive e2e test for the build-db process, which
was until now rather untested.

This adds a new helper that enables comparing and generating
an entire directory with test data. It's used to generate and
compare the full DB directory. Everything under `db/e2e_expected`
is auto-generated and shows the status quo.

The inputs to the new tests span various examples of extractable
and non-extractable expressions. Some of which are debatable,
but this is up to future improvements.

Bug: 389069288
Change-Id: I49785d0615b901e99b1e82b825fde5a52f2fe2fc
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6368714
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Stephen Röttger <sroettger@google.com>
Cr-Commit-Position: refs/heads/main@{#99325}