[fuchsia] Check alignment of AllocatePages when necessary
Fuchsia has a maximum alignment of 2^32. CppGC, however, tries to
allocate the HeapCage with an alignment of 2^35, and also provides a
placement hint fot that allocation. The validation of the alignment
parameter of that allocation therefore fails. However, since the
allocation provides a placement hint, the hint can be used without
alignment validation.

With this CL the alignment gets validated only when necessary, which
means only when the placement hint is not supposed to be used, or when
the placement hint did not work out.

Bug: 42203693
Change-Id: Ie02f3358237917704ae5d7eecc1f230c32b39e2a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6357399
Reviewed-by: Omer Katz <omerkatz@chromium.org>
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Auto-Submit: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99332}