[execution, arm64] Fix fast c calls in arm64 simulator with profiler
In simulator builds the Stp instruction which is used on arm64
to store both the fast_c_call_caller_fp and fast_c_call_caller_pc at
once is not signal-safe. A signal can happen between storing the
FP first and only then the PC field.

This is an issue during profiling (which uses signals) because the
profiler assumes that once the FP is set, the PC also always needs
to be set.

This CL solves this by reordering the FP and PC fields in the isolate.
With this change having a signal in-between the two stores isn't an
issue anymore because the PC field is always set before the FP now.

Bug: 402310230
Change-Id: I404cfe814ba0365f8764d538964a3786973bab33
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6368520
Reviewed-by: Andreas Haas <ahaas@chromium.org>
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99341}