[regexp] Fix common prefix extraction when reading backwards
We extract common prefixes out of disjunctions (e.g. ab|ac|ad becomes
a(?:b|c|d). To do so alternatives are stable-sorted, comparing the first
character only. In general the order of alternatives matter, as they
have to be evaluated left to right. As we only compare on the first
char, this is usually fine as alternatives with the same first char stay
in the same order and for alternatives with different first chars only
one of them can match.
However this optimization (specifically stable-sorting on the first
char) is not correct if we read backwards (e.g. for look-behind
assertions). To fix this issue, we extract common prefixes when we read
forward, and common suffixes when reading backward.

Fixed: 388290816
Change-Id: I86fb87462f6f0affd14a08904f924be77e683fce
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6372663
Commit-Queue: Patrick Thier <pthier@chromium.org>
Reviewed-by: Jakob Linke <jgruber@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99339}