[serializer] Promote FunctionTemplateInfo to RO space when possible
This CL adds
v8::FunctionTemplate::SealAndPrepareForPromotionToReadOnly() which
is supposed to be called by embedder when the function template is
ready and is not going to be modified.

In case serializer fails to promote such a template to read only
space a CHECK failure will be triggered.

Bug: 399393891
Change-Id: Ia8f205a84b5ae427932ca6e9fd25bd04ec596908
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6352058
Reviewed-by: Jakob Linke <jgruber@chromium.org>
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Igor Sheludko <ishell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99353}