[test] Fix TestExperimentalNotEnabled for explicit flags
TestExperimentalNotEnabled checks that some specific configurations
enable the "--experimental" flag, which is otherwise disabled. This
test fails if an explicit experimental flag is used when the unit test
is invoked. This CL fixes the test by checking that, in such cases,
the experimental flag remains enabled. The change allows us to run
unit tests in experimental variants on the bots, as well as to run
unit tests locally with experimental flags.

Change-Id: I8a1bb805df49cbd2ede0b769597c206c7d57f027
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6375723
Reviewed-by: Omer Katz <omerkatz@chromium.org>
Reviewed-by: Michael Achenbach <machenbach@chromium.org>
Commit-Queue: Nikolaos Papaspyrou <nikolaos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99368}