[json] Stringify: Reserve buffer capacity upfront
To avoid repeated checks if the current character(s) fit into the
current buffer segment, we reserve enough space upfront for multiple
appends if we know the size statically upfront.
E.g. when appending an object key, we need to (optionally) append a
comma, a double quote, object key itself, a double quote, and a colon.
Previously we checked the capacity for each individual added character
literal and the object key (up to 5 checks), which are now folded into a
single check.
For this logic to work, we now have variable sized segments in the
buffer (segments can shrink if we reserve space that won't fit in the
current segment, and we need to allocate a new one).

Drive-by: AppendStringChecked* variants are renamed to AppendString*.
The "Checked" previously indicated that we check for characters that
need escaping, but this clashes with the newly introduced naming scheme
of "*Unchecked" for methods that write into the buffer without capacity
checks.

Bug: 380044242
Change-Id: I9c4d25807f25ad50fc5781116447f7db7bbe045a
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6375724
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Patrick Thier <pthier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99369}