[heap] Use Size() for implementing AllocatedSinceLastGC()
We can compute the number of allocated bytes since the last GC by
comparing Size() calls as well. This is quite a bit simpler.

AllocateSinceLastGC() also returned a wrong value when a
"parkable buffer" was on the page with the age mark and that page
got moved to the end of the memory_chunk_list_ in SemiSpace.

Bug: 391129951
Change-Id: I8d3c9642cd9acd80156ffb72373309002018543f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6375292
Commit-Queue: Dominik Inführ <dinfuehr@chromium.org>
Reviewed-by: Omer Katz <omerkatz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99365}