[wasm] Use JSAny for stricter type checks
We had issues with leaking the termination exception to JS when taking a
pending exception from the isolate and setting it as the promise
rejection value.
This CL makes the C++ types more strict by requiring `JSAny` for the
error reason instead of `Object`. This requires a cast at callers which
do not already have a static subtype of `JSAny`, and thus makes it more
obvious that certain types (like holes) have to be excluded.

R=jkummerow@chromium.org

Bug: 403308742
Change-Id: Ic764f9db4e6979174eedd5cd1f1e43bb9df32ab4
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6375484
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99360}