[gdbinit] Detect a few more interesting frames
gdbinit looks through the first handful of frames and if it finds an
"interesting" entry point it selects the parent frame of that. This
often brings you directly to the DCHECK or FATAL which fails.
This CL extends this list by a few more frames, in particular around
failed ApiChecks.

R=jkummerow@chromium.org

No-Try: true
Change-Id: Ifbf04eb261bf28f330c3b12988ffd20a84424bde
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6377323
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99374}