[wasm][jspi] Run the unwinder on the central stack
After switching to the central stack to perform a runtime call, only
switch back to the secondary stack if there was no exception.
If there is an exception, we unwind the stack, so update the
stack-switching data based on the target handler instead.

This should resolve the `IsOnStack` failures from:
https://chromium-review.googlesource.com/c/v8/v8/+/6087918

since the unwinder now runs from the central stack.

R=jkummerow@chromium.org
CC=​​nikolaos@chromium.org

Bug: 42203436
Change-Id: Ic9bd95cca46d93ac4ca3c577ceee99a89b5abe62
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6373484
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99372}