[api] Fix down-casts of MaybeLocal
Down-casts of MaybeLocal were completely broken. The line
`return MaybeLocal<T>(that.local_);` called the
`MaybeLocal<Local<S>>` constructor which then called `Local<Local<S>>`
for up-casts, but for down-casts this constructor cannot be used because
of the requires-clause, so it instead tried to called
`LocalBase<LocalBase<S>>`, which is protected and leads to a compiler
error.

This CL fixes that by just using `Cast` on the `Local`, which is
correctly implemented also for empty `Local`s, and then converts that
back into a `MaybeLocal`.
It also adds implicit up-casts on `MaybeLocal`, and adds tests for all
different kinds of casts.

R=mlippautz@chromium.org

Change-Id: I342e84bf594a728889f775abf16b9cfe4bf82d84
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6381814
Reviewed-by: Michael Lippautz <mlippautz@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99387}