[wasm][jspi] Cleanup UpdateCentralStackInfo
With JSPI, resuming a stack always switches from the central stack to a
secondary stack, and returning/suspending always switches from a
secondary stack to the central stack.
Based on this, we can simplify the logic for updating the central stack
data on stack switch.

R=jkummerow@chromium.org

Bug: 404285918
Change-Id: Icbcb21bfa6728ffd20de7cf3b7867f2899fa3d7e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6381013
Commit-Queue: Thibaud Michaud <thibaudm@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99376}