[heap] Add IsolateScope for IncrementalMarkingJob
Same as https://crrev.com/c/6280810, but for incremental marking. We
need to set the current isolate in thread-local storage so we can access
pointer tables (in particular the trusted pointer table) from GC.

Also add a DCHECK that we only allocate in the current isolate (because
allocation can trigger GC, but we want to fail early and not randomly on
GC).
This also needs a few fixes in tests which allocated without having
entered the isolate.

R=dinfuehr@chromium.org
CC=​mlippautz@chromium.org

Bug: 405047989, 405172561, 396607238
Change-Id: Iaa52663faa6f2900a8d561c4cb4d77c728e47bcc
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6381773
Reviewed-by: Dominik Inführ <dinfuehr@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99378}