[js_fuzzer] Update Babel traverse
The upgrade is required for issue 406389895. In Babel 7.6.0 however,
a bug was fixed that changes the behavior to fetch available
identifiers. Though, it's correct now, the previous version might
have been more beneficial for fuzzing.

We'll look into a workaround to extend the available identifiers
in a follow up.

The expectation changes are all due to the fixed bug. In some
examples, there are expectation changes due to the pseudo-random
test flow.

Bug: 389069288, 406389895
Change-Id: I0a95be796f232e9172db973342efb744b1704f66
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6383037
Commit-Queue: Michael Achenbach <machenbach@chromium.org>
Reviewed-by: Stephen Röttger <sroettger@google.com>
Cr-Commit-Position: refs/heads/main@{#99512}