[heap] Check that we only allocate in the current isolate
Add a DCHECK that we only allocate in the current isolate (because
allocation can trigger GC, which requires the isolate to be set; the
DCHECK ensures that we fail early and not randomly on GC).
This also needs a few fixes in tests which allocated without having
entered the isolate.

R=dinfuehr@chromium.org
CC=​mlippautz@chromium.org

Bug: 396607238
Change-Id: Ibbe81b498d1948c2f1c6a85126da703b3d1fb5bd
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6383258
Reviewed-by: Dominik Inführ <dinfuehr@chromium.org>
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99439}