[ic][test] Fix StubCache test and add DCHECK for correct cage
The handlers stored in the StubCache data structure must be either
Smi or heap-allocated objects living in the regular cage (when pointer
compression is enabled), because they are stored in the cache as
compressed pointers and StubCache::Get() implicitly uses the regular
cage base. This CL adds a DCHECK to StubCache::Set() enforcing this.
It also fixes the cctest TryProbeStubCache, which until now stored
handlers that lived in the code cage, and adds one more basic test.

Change-Id: If2c6e7c0b361b2c9acadf86bb383d8cd4ca13705
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6387118
Reviewed-by: Igor Sheludko <ishell@chromium.org>
Commit-Queue: Nikolaos Papaspyrou <nikolaos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99411}