[wasm] Skip space for far jump table if not needed
On platforms where no far jumps are needed, and thus also no far jump
table, take this into account when computing the code space reservation
size.

This results in the reservation size sometimes being 0, which again
requires some follow-up changes (to not try to allocate a virtual memory
of size 0).
To ensure that we actually never try to allocate a code space for
modules without any declared functions, we add a number of DCHECKs, and
fix test code and the wrapper cache to not hit those DCHECKs.

R=jkummerow@chromium.org

Bug: 404568041
Change-Id: I3607cbba20ad2c13a47afae5c74a15d654d4583e
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6381741
Commit-Queue: Clemens Backes <clemensb@chromium.org>
Reviewed-by: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99418}