[sandbox] Fix size of filler object for clearing preparse data
This CL fixes SharedFunctionInfo::ClearPreparseData, which assumed
that the object to be trimmed was an UncompiledDataWithPreparseData,
so that it calculates correctly the size of the filler object when
the object to be trimmed comes from a subtype, such as
UncompiledDataWithPreparseDataAndJob. This has been known to cause
test mjsunit/regress/regress-350779648 to flake for stress_snapshot.
The test is re-enabled by this CL.

Change-Id: I2441d05a5dcb6aee5a835b9afdfe4c2dc001cc15
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6394980
Reviewed-by: Dominik Inführ <dinfuehr@chromium.org>
Commit-Queue: Nikolaos Papaspyrou <nikolaos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99478}