Revert: [maglev] Don't bail out if we don't change a const field
Fields are always updated to be mutable once they are stored to.
Thus this case is actually unreachable. Either the field is already
mutable. Or, we never executed this store before -- in which case
we lack the necessary feedack to end up in `TryBuildStoreField`.

See
https://chromium-review.googlesource.com/c/v8/v8/+/629939

Change-Id: I01e4675887a1c985d58769f134d3915f18f59f2f
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6396661
Reviewed-by: Olivier Flückiger <olivf@chromium.org>
Commit-Queue: Victor Gomes <victorgomes@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99480}