[Trace profile] Point from trace id to node id in traced samples
Instead of sample index to trace id. This makes the manually collected
stack sample independent of the order in which it appears in the CPU
profile, simplifying its parsing by the consumer. In DevTools, this
allows to remove a level of indirection, since we were obtaining the
node id for a sample trace id based on the index the sample occupied in
the profile. Now we can take it directly from the trace id data. (see
https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6394268)

Bug: 403628855
Change-Id: I641f50ff0e19cbbcdc7a281e662cf5737b23860c
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/6397026
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#99495}